+++ lib/os/Kconfig
@@ -214,5 +214,6 @@ rsource "Kconfig.atomics"
 rsource "Kconfig.cbprintf"
 rsource "zvfs/Kconfig"
 rsource "cpu_load_metric/Kconfig"
+rsource "cpu_load/Kconfig"
 
 endmenu
