+++ lib/os/cpu_load/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright (c) 2024, Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources(cpu_load.c)
