+++ lib/os/cpu_load/Kconfig
@@ -1,0 +1,38 @@
+# Copyright (c) 2024, Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: Apache-2.0
+
+config CPU_LOAD
+	bool "CPU load measurement"
+	depends on CPU_CORTEX_M || RISCV || CPU_CORTEX_A || XTENSA
+	depends on !SMP
+	select SYS_IDLE_HOOKS
+	help
+	  Track CPU load by measuring the time spent in the idle loop using the
+	  architecture idle hooks. Does not depend on the tracing subsystem.
+
+if CPU_LOAD
+
+# Workaround for not being able to have commas in macro arguments
+DT_CHOSEN_Z_CPU_LOAD_COUNTER := zephyr,cpu-load-counter
+
+config CPU_LOAD_USE_COUNTER
+	bool "Use counter"
+	depends on $(dt_chosen_enabled,$(DT_CHOSEN_Z_CPU_LOAD_COUNTER))
+	default y
+	select COUNTER
+	help
+	  Use counter for tracking CPU idle time.
+
+config CPU_LOAD_LOG_PERIODICALLY
+	int "Report period (in milliseconds)"
+	depends on LOG
+	default 0
+	help
+	  Specifies how often CPU load shall be logged. 0 means that there is no logging.
+
+module = CPU_LOAD
+module-str = cpu_load
+source "subsys/logging/Kconfig.template.log_config"
+
+endif # CPU_LOAD
