+++ modules/lvgl/lvgl_zephyr_osal.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/debug/cpu_load.h>
+#include <zephyr/sys/cpu_load.h>
 #include <lvgl.h>
 #include <lvgl_private.h>
