+++ subsys/tracing/tracing_none.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #include <zephyr/tracing/tracing.h>
-#include <zephyr/debug/cpu_load.h>
+#include <zephyr/sys/cpu_load.h>
 
 __weak void sys_trace_isr_enter(void) {}
